Added the copyCat.java project onto the repository
The copyCat.java is called copy cat simply because it is a project derived from the "game.java" class i created . 
Author: Terrell Martin
Date: 07/04/2018
Purpose: The purpose of the copyCat.java is to gain more knowledge of OOP and designing object onto the specified panels. Also, it is meant to be different in it's very own way by designing a RPG like feel.